requirements.txt
authy
django-phonenumber-field
django-countries
-e git+https://github.com/rosscdh/django-jsonfield.git@lawpal#egg=django-jsonfield